refactor(layout): shift siteviewlayout into layouts/layouts
